python_client/requirements.txt
pulsar-client==2.7.0